[#162986280] MainActivity 관련 코드 정리
