+++ app/src/main/java/com/playgilround/schedule/client/activity/MainActivity.java
@@ -1,0 +1,52 @@
+package com.playgilround.schedule.client.activity;
+
+import android.Manifest;
+import android.content.Intent;
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+import android.view.View;
+
+import com.gun0912.tedpermission.PermissionListener;
+import com.gun0912.tedpermission.TedPermission;
+import com.playgilround.schedule.client.R;
+
+import java.util.ArrayList;
+
+public class MainActivity extends AppCompatActivity {
+
+    private String[] PERMISSION_STORAGE = {
+            Manifest.permission.ACCESS_FINE_LOCATION,
+            Manifest.permission.ACCESS_COARSE_LOCATION,
+            Manifest.permission.INTERNET
+    };
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+
+        TedPermission.with(this)
+                .setPermissionListener(new PermissionListener() {
+                    @Override
+                    public void onPermissionGranted() {
+                        
+                    }
+
+                    @Override
+                    public void onPermissionDenied(ArrayList<String> deniedPermissions) {
+
+                    }
+                })
+                .setDeniedMessage(getString(R.string.message_check_permission))
+                .setPermissions(PERMISSION_STORAGE)
+                .check();
+
+        findViewById(R.id.button_calendar).setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                startActivity(new Intent(MainActivity.this, ScheCalendarActivity.class));
+            }
+        });
+    }
+}
+
