+++ app/src/main/res/values-ko/strings.xml
@@ -14,4 +14,7 @@
     <string name="hint_schedule">스케줄을 적어주세요..</string>
     <string name="hint_schedule_contents">내용을 적어주세요..</string>
 
+    <!-- Message -->
+    <string name="message_check_permission">권한 요청을 해주세요.</string>
+
 </resources>
